-working on #125 -minimum price set (0.5€). Closes #126 -load images on bootstrapping (dev-postgresql)

+++ src/main/java/misrraimsp/uned/pfg/firstmarket/config/dev/sqlBuilder/Random.java
@@ -6,7 +6,7 @@ public class Random {
         return (int) (low_limit + Math.round(Math.random() * (up_limit - low_limit)));
     }
 
-    public double getContinuousRandomNumber(int low_limit, int up_limit){
+    public double getContinuousRandomNumber(double low_limit, double up_limit){
         return low_limit + Math.random() * (up_limit - low_limit);
     }
 }
